// src/services/-admin-inventory-adjustments-recent-get.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Shopping Platform API
 * Production-grade e-commerce backend
 * OpenAPI spec version: 1.0
 */
export interface UtilsResponse {
  code?: number;
  error?: string;
  message?: string;
  success?: boolean;
}

export interface DtoInventoryAdjustmentResponse {
  actor_name?: string;
  adjustment_type?: string;
  created_at?: string;
  id?: number;
  note?: string;
  product_id?: number;
  product_name?: string;
  product_sku?: string;
  quantity_after?: number;
  quantity_before?: number;
  quantity_delta?: number;
  reference_id?: number;
  reference_type?: string;
}

export type GetAdminInventoryAdjustmentsRecentParams = {
/**
 * Max rows
 */
limit?: number;
};

export type GetAdminInventoryAdjustmentsRecent200 = UtilsResponse & {
  data?: DtoInventoryAdjustmentResponse[];
};